!T: fix warning on x86 build

+++ core/rt/json.cpp
@@ -519,7 +519,7 @@ void JsonBeautified::_Beautify(const rt::String_Ref& json_string, bool newline, 
 		return;
 	}
 
-	if(doc.GetLength() < line_remain)
+	if(((int)doc.GetLength()) < line_remain)
 	{
 		int seps = JsonObject::_count_seps(doc);
 		if(seps == 0 || (doc.GetLength() < (seps+1)*16 && seps<=3))
